requirement.txt
pandas==1.4.1
numpy==1.22.2
scipy==1.7.3
ipython==8.0.1
google-api-python-client==2.37.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-storage==2.1.0
google-cloud-bigquery==2.34.0
tensorflow==2.8.0
spacy==3.2.0
sentence-transformers==2.2.0 # not available in docker image 
gensim==4.1.2
nltk==3.7
jellyfish==0.9.0
scikit-learn==1.0.1
xgboost==1.5.2 # not available in docker image 
PyYAML==6.0
